requirements.txt
anyjson==0.3.3
APScheduler==3.7.0
asgiref==3.3.1
autopep8==1.5.6
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery~=5.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-apscheduler==0.5.2
django-choices==1.7.1
django-redis==4.12.1
django-silk==4.1.0
django-windows-tools==0.2.1
djangorestframework==3.12.2
dnspython==1.16.0
drf-yasg==1.20.0
eventlet==0.30.2
factory-boy==3.2.0
Faker==6.5.0
gevent==21.1.2
gprof2dot==2021.2.21
greenlet==1.0.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mock==4.0.3
mysqlclient==2.0.3
packaging==20.9
Pillow==8.1.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.15.90
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
gunicorn==20.0.4
black==21.5b1
pre-commit==2.12.1
profanity-filter==1.3.3
better-profanity==0.7.0